made changes as per the code review
