--- app/helpers/api/v1/contacts_helper.rb
@@ -1,4 +1,0 @@
-# frozen_string_literal: true
-
-module Api::V1::ContactsHelper
-end
