+++ app/javascript/src/components/Dashboard/Contacts/DeleteAlert.jsx
@@ -1,21 +1,18 @@
-/* eslint-disable arrow-body-style */
 import React from "react";
 
 import { Alert, Toastr } from "neetoui";
 
-const DeleteAlert = ({ onClose }) => {
-  return (
-    <Alert
-      isOpen
-      message="Are you sure you want to continue? This cannot be undone."
-      title="Delete Contact"
-      onClose={onClose}
-      onSubmit={() => {
-        Toastr.success("Contact deleted successfully.");
-        onClose();
-      }}
-    />
-  );
-};
+const DeleteAlert = ({ onClose }) => (
+  <Alert
+    isOpen
+    message="Are you sure you want to continue? This cannot be undone."
+    title="Delete Contact"
+    onClose={onClose}
+    onSubmit={() => {
+      Toastr.success("Contact deleted successfully.");
+      onClose();
+    }}
+  />
+);
 
 export default DeleteAlert;
