+++ app/javascript/src/components/Dashboard/Contacts/Pane/Form.jsx
@@ -70,6 +70,7 @@ const NoteForm = ({ onClose, contact }) => {
               label="Cancel"
               size="large"
               style="text"
+              type="reset"
               onClick={onClose}
             />
           </Pane.Footer>
