+++ app/javascript/src/components/Dashboard/Contacts/Table/index.jsx
@@ -1,23 +1,21 @@
-/* eslint-disable */
 import React from "react";
-import { Table as NeetoTable } from "neetoui";
-import { CONTACTS_LIST } from "../constants";
-import { contactsTableColumnData } from "./TableData";
 
-const Table = ({ setShowDeleteAlert }) => {
-  return (
-    <NeetoTable
-      columnData={contactsTableColumnData(setShowDeleteAlert)}
-      currentPageNumber={1}
-      defaultPageSize={10}
-      handlePageChange={function noRefCheck() {}}
-      onRowClick={function noRefCheck() {}}
-      onRowSelect={function noRefCheck() {}}
-      rowData={CONTACTS_LIST}
-      pagination
-      allowRowClick
-    />
-  );
-};
+import { Table as NeetoTable } from "neetoui";
+
+import { CONTACTS_LIST, contactsTableColumnData } from "../constants";
+
+const Table = ({ setShowDeleteAlert }) => (
+  <NeetoTable
+    allowRowClick
+    pagination
+    columnData={contactsTableColumnData(setShowDeleteAlert)}
+    currentPageNumber={1}
+    defaultPageSize={10}
+    handlePageChange={() => {}}
+    rowData={CONTACTS_LIST}
+    onRowClick={() => {}}
+    onRowSelect={() => {}}
+  />
+);
 
 export default Table;
