+++ app/javascript/src/components/Dashboard/Notes/Pane/Form.jsx
@@ -82,6 +82,7 @@ const NoteForm = ({ onClose, refetch, note, isEdit }) => {
               label="Cancel"
               size="large"
               style="text"
+              type="reset"
               onClick={onClose}
             />
           </Pane.Footer>
