--- test/controllers/api/v1/contacts_controller_test.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-require "test_helper"
-
-class Api::V1::ContactsControllerTest < ActionDispatch::IntegrationTest
-  # test "the truth" do
-  #   assert true
-  # end
-end
